fix: resolve repository_settings from inside a worktree (#84)
When `gwq add` was run from inside an existing worktree,
`repository_settings` (setup_commands, copy_files) were not applied
because `os.Getwd()` returned the worktree path instead of the main
repository root, causing the config match to fail.

Use `git rev-parse --git-common-dir` to always resolve the main
repository root regardless of whether the caller is in the main repo
or a worktree. Also resolve symlinks on both the git-derived path and
the config-defined repository path for consistent comparison.

Additionally:
- Fix ListWorktrees IsMain detection from worktree context
- Split git.go into git_branch.go, git_repo.go, git_worktree.go
- Extract runPostWorktreeSetup into post_setup.go
- Remove backward-compat Run() alias (unified to RunCommand)

Closes #84
