+++ internal/git/git.go
@@ -7,11 +7,7 @@ import (
 	"fmt"
 	"os"
 	"os/exec"
-	"path/filepath"
 	"strings"
-	"time"
-
-	"github.com/d-kuro/gwq/pkg/models"
 )
 
 // Git provides Git command operations.
@@ -35,304 +31,13 @@ func NewFromCwd() (*Git, error) {
 	return New(cwd), nil
 }
 
-// ListWorktrees returns a list of all worktrees in the repository.
-func (g *Git) ListWorktrees() ([]models.Worktree, error) {
-	output, err := g.run("worktree", "list", "--porcelain")
-	if err != nil {
-		return nil, fmt.Errorf("failed to list worktrees: %w", err)
-	}
-
-	var worktrees []models.Worktree
-	lines := strings.Split(strings.TrimSpace(output), "\n")
-
-	for i := 0; i < len(lines); i++ {
-		if after, ok := strings.CutPrefix(lines[i], "worktree "); ok {
-			path := after
-
-			var branch, commitHash string
-			isMain := false
-
-			for j := i + 1; j < len(lines) && !strings.HasPrefix(lines[j], "worktree "); j++ {
-				if after, ok := strings.CutPrefix(lines[j], "branch "); ok {
-					branch = after
-					// Remove refs/heads/ prefix if present
-					branch = strings.TrimPrefix(branch, "refs/heads/")
-				} else if after, ok := strings.CutPrefix(lines[j], "HEAD "); ok {
-					commitHash = after
-				} else if strings.HasPrefix(lines[j], "bare") {
-					continue
-				}
-				i = j
-			}
-
-			if branch == "" {
-				branch = g.getCurrentBranch(path)
-			}
-
-			info, err := os.Stat(path)
-			var createdAt time.Time
-			if err == nil {
-				createdAt = info.ModTime()
-			}
-
-			worktrees = append(worktrees, models.Worktree{
-				Path:       path,
-				Branch:     branch,
-				CommitHash: commitHash,
-				IsMain:     isMain,
-				CreatedAt:  createdAt,
-			})
-		}
-	}
-
-	if len(worktrees) > 0 {
-		mainDir, err := g.getMainWorktreeDir()
-		if err == nil {
-			for i := range worktrees {
-				if worktrees[i].Path == mainDir {
-					worktrees[i].IsMain = true
-					break
-				}
-			}
-		}
-	}
-
-	return worktrees, nil
-}
-
-// AddWorktree creates a new worktree.
-func (g *Git) AddWorktree(path, branch string, createBranch bool) error {
-	args := []string{"worktree", "add"}
-
-	if createBranch {
-		args = append(args, "-b", branch, path)
-	} else {
-		args = append(args, path, branch)
-	}
-
-	if _, err := g.run(args...); err != nil {
-		return fmt.Errorf("failed to add worktree: %w", err)
-	}
-
-	return nil
-}
-
-// AddWorktreeFromBase creates a new worktree with a branch from a specific base branch.
-func (g *Git) AddWorktreeFromBase(path, branch, baseBranch string) error {
-	args := []string{"worktree", "add", "-b", branch, path}
-
-	if baseBranch != "" {
-		args = append(args, baseBranch)
-	}
-
-	if _, err := g.run(args...); err != nil {
-		return fmt.Errorf("failed to add worktree from base branch %s: %w", baseBranch, err)
-	}
-
-	return nil
-}
-
-// RemoveWorktree removes a worktree.
-func (g *Git) RemoveWorktree(path string, force bool) error {
-	args := []string{"worktree", "remove"}
-	if force {
-		args = append(args, "--force")
-	}
-	args = append(args, path)
-
-	if _, err := g.run(args...); err != nil {
-		return fmt.Errorf("failed to remove worktree: %w", err)
-	}
-
-	return nil
-}
-
-// DeleteBranch deletes a branch.
-func (g *Git) DeleteBranch(branch string, force bool) error {
-	args := []string{"branch"}
-	if force {
-		args = append(args, "-D")
-	} else {
-		args = append(args, "-d")
-	}
-	args = append(args, branch)
-
-	if _, err := g.run(args...); err != nil {
-		return fmt.Errorf("failed to delete branch %s: %w", branch, err)
-	}
-
-	return nil
-}
-
-// PruneWorktrees removes worktree information for deleted directories.
-func (g *Git) PruneWorktrees() error {
-	if _, err := g.run("worktree", "prune"); err != nil {
-		return fmt.Errorf("failed to prune worktrees: %w", err)
-	}
-	return nil
-}
-
-// ListBranches returns a list of all branches.
-func (g *Git) ListBranches(includeRemote bool) ([]models.Branch, error) {
-	args := []string{"branch", "-v", "--format=%(refname:short)|%(HEAD)|%(committerdate:iso)|%(objectname)|%(subject)|%(authorname)"}
-	if includeRemote {
-		args = append(args, "-a")
-	}
-
-	output, err := g.run(args...)
-	if err != nil {
-		return nil, fmt.Errorf("failed to list branches: %w", err)
-	}
-
-	var branches []models.Branch
-	lines := strings.SplitSeq(strings.TrimSpace(output), "\n")
-
-	for line := range lines {
-		if line == "" {
-			continue
-		}
-
-		parts := strings.Split(line, "|")
-		if len(parts) < 6 {
-			continue
-		}
-
-		name := parts[0]
-		isCurrent := parts[1] == "*"
-		dateStr := parts[2]
-		hash := parts[3]
-		message := parts[4]
-		author := parts[5]
-
-		isRemote := strings.HasPrefix(name, "remotes/")
-		if isRemote {
-			name = strings.TrimPrefix(name, "remotes/")
-		}
-
-		date, _ := time.Parse("2006-01-02 15:04:05 -0700", dateStr)
-
-		branches = append(branches, models.Branch{
-			Name:      name,
-			IsCurrent: isCurrent,
-			IsRemote:  isRemote,
-			LastCommit: models.CommitInfo{
-				Hash:    hash,
-				Message: message,
-				Author:  author,
-				Date:    date,
-			},
-		})
-	}
-
-	return branches, nil
-}
-
-// GetRepositoryName returns the name of the repository.
-func (g *Git) GetRepositoryName() (string, error) {
-	rootDir, err := g.getRootDir()
-	if err != nil {
-		return "", err
-	}
-	return filepath.Base(rootDir), nil
-}
-
-// GetRepositoryPath returns the root path of the git repository.
-func (g *Git) GetRepositoryPath() (string, error) {
-	return g.getRootDir()
-}
-
-// GetRecentCommits returns recent commits for a specific path.
-func (g *Git) GetRecentCommits(path string, limit int) ([]models.CommitInfo, error) {
-	oldWorkDir := g.workDir
-	g.workDir = path
-	defer func() { g.workDir = oldWorkDir }()
-
-	args := []string{"log", fmt.Sprintf("-%d", limit), "--pretty=format:%H|%s|%an|%ai"}
-	output, err := g.run(args...)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get recent commits: %w", err)
-	}
-
-	var commits []models.CommitInfo
-	lines := strings.SplitSeq(strings.TrimSpace(output), "\n")
-
-	for line := range lines {
-		if line == "" {
-			continue
-		}
-
-		parts := strings.Split(line, "|")
-		if len(parts) < 4 {
-			continue
-		}
-
-		date, _ := time.Parse("2006-01-02 15:04:05 -0700", parts[3])
-
-		commits = append(commits, models.CommitInfo{
-			Hash:    parts[0],
-			Message: parts[1],
-			Author:  parts[2],
-			Date:    date,
-		})
-	}
-
-	return commits, nil
-}
-
-// getCurrentBranch returns the current branch name for a specific worktree.
-func (g *Git) getCurrentBranch(worktreePath string) string {
-	oldWorkDir := g.workDir
-	g.workDir = worktreePath
-	defer func() { g.workDir = oldWorkDir }()
-
-	output, err := g.run("rev-parse", "--abbrev-ref", "HEAD")
-	if err != nil {
-		return ""
-	}
-	return strings.TrimSpace(output)
-}
-
-// getMainWorktreeDir returns the main worktree directory.
-func (g *Git) getMainWorktreeDir() (string, error) {
-	return g.getRootDir()
-}
-
-// getRootDir returns the repository root directory.
-func (g *Git) getRootDir() (string, error) {
-	output, err := g.run("rev-parse", "--show-toplevel")
-	if err != nil {
-		return "", fmt.Errorf("failed to get repository root: %w", err)
-	}
-	return strings.TrimSpace(output), nil
-}
-
-// GetRepositoryURL returns the remote origin URL of the repository.
-func (g *Git) GetRepositoryURL() (string, error) {
-	output, err := g.run("remote", "get-url", "origin")
-	if err != nil {
-		return "", fmt.Errorf("failed to get repository URL: %w", err)
-	}
-	return strings.TrimSpace(output), nil
-}
-
 // RunCommand executes a git command with the provided arguments and returns the output.
-// This is the primary public interface for running arbitrary git commands.
 // The command is executed in the Git instance's working directory if set.
-// Returns the command output as a string, or an error if the command fails.
 func (g *Git) RunCommand(args ...string) (string, error) {
 	return g.run(args...)
 }
 
-// Run is an alias for RunCommand maintained for backward compatibility.
-// New code should prefer using RunCommand for clarity.
-func (g *Git) Run(args ...string) (string, error) {
-	return g.run(args...)
-}
-
 // RunWithContext executes a git command with context support for cancellation and timeout.
-// The command will be terminated if the context is cancelled or times out.
-// This is useful for implementing timeouts or graceful shutdowns.
-// Returns the command output as a string, or an error if the command fails or is cancelled.
 func (g *Git) RunWithContext(ctx context.Context, args ...string) (string, error) {
 	return g.runWithContext(ctx, args...)
 }
@@ -367,7 +72,6 @@ func (g *Git) runWithContext(ctx context.Context, args ...string) (string, error
 	cmd.Stderr = &stderr
 
 	if err := cmd.Run(); err != nil {
-		// Check if context was cancelled or timed out
 		if ctx.Err() != nil {
 			return "", fmt.Errorf("git %s: %w", strings.Join(args, " "), ctx.Err())
 		}
